rafactor(coffeeBannerGenerator): renamed components to PascalCase, fixed img source in clipboard, disabled custom input, removed unnecessary tailwind classes and moved radio component to components folder.
