+++ src/components/radio-input/RadioInput.tsx
@@ -15,16 +15,16 @@ interface RadioInputProps {
 const RadioInput = ({ value, disabled, checked, divClassName, handleChange, labelType='label', label, name, width }: RadioInputProps) => {
   const [inputVal, setInputVal] = useState(value);
 
-  const handleInput = (value) => {
-    setInputVal(value);
-    handleChange(value);
+  const handleInput = (value: string) => {
+    setInputVal(value+'px');
+    handleChange(value+'px');
   }
 
   return (
     <div className={`flex items-center gap-1 ${divClassName}`}>
         <input disabled={disabled}  type="radio" id={value.toLocaleString()} name={name} value={value} checked={width ? inputVal == width : checked} onChange={(e) => handleChange(e.target.value)} />
         {
-            labelType == 'input' ? <input value={inputVal} onChange={(e) => handleInput(e.target.value)} className={`text-end px-1 w-[60px] border-lightGray border-2 ${(width ? inputVal == width : checked) ? '' : 'text-lightGray'}`} /> : <label className={checked ? '' : 'text-lightGray'}>{label}</label> 
+            labelType == 'input' ? <input disabled={disabled} type='text' value={inputVal} onChange={(e) => handleInput(e.target.value)} className={`text-end px-1 w-[60px] border-lightGray border-2 ${(width ? inputVal == width : checked) ? '' : 'text-lightGray'}`} /> : <label className={checked ? '' : 'text-lightGray'}>{label}</label> 
         }
     </div>
   );
