+++ src/components/step-card/StepCard.tsx
@@ -6,7 +6,7 @@ interface StepCardProps {
 
 function StepCard({title ,stepNumber, children } : StepCardProps) {
   return (
-    <div className='w-full rounded-lg shadow-lg'>
+    <div className='w-full rounded-lg shadow-lg my-2'>
         <div className='bg-darkGray m-0 rounded-t-lg py-2 px-4 font-bold text-xl text-white'>
             <h3>{stepNumber}. {title}</h3>
         </div>
